# internal/handler/openapi.yaml
openapi: 3.1.0
info:
  title: SN360-ES Service API
  version: "v2.0"
  description: |
    REST surface for the SN360-ES (Email Security) control plane.

    The API is divided into:
      * `/v1/banner/*` — one-click user feedback from rendered banners
      * `/v1/quarantine/*` — quarantine release approval workflow
      * `/v1/education/*` — micro-lesson and simulation tracking
      * `/v1/escalation/*` — SOC escalation ticket handling
      * `/v1/dashboard/*` — aggregated tenant dashboards
      * `/v1/predict/*` — pre-send and pre-open add-in scoring
      * `/v1/onboarding/*` — OAuth consent flow for tenant onboarding

    All request and response bodies are JSON. Authentication is via
    short-lived signed action tokens or tenant API keys depending on the
    endpoint.
  contact:
    name: SN360-ES Maintainers
    url: https://github.com/kennguy3n/sn360-es
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://sn360-es.dev.uney.com
    description: Dev cluster
...
paths:
...
  /v1/banner/action:
    post:
      tags: [banner]
      summary: Submit one-click banner feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerActionRequest"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BannerActionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/quarantine/release:
    post:
      tags: [quarantine]
      summary: Release a quarantined message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuarantineReleaseRequest"
      responses:
        "200":
          description: Release outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuarantineReleaseResponse"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/education/lesson/{category}:
    get:
      tags: [education]
      summary: Fetch a micro-lesson
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
            description: "One of the SN360-ES risk categories (e.g. LIKELY_PHISHING, BEC_IMPERSONATION, SUSPICIOUS_URL). See README for the full list."
        - in: query
          name: locale
          schema:
            type: string
            default: en
      responses:
        "200":
          description: Lesson body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroLesson"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/escalation/resolve:
    post:
      tags: [escalation]
      summary: Resolve a SOC escalation ticket
      description: |
        Records the SecOps outcome on an existing escalation ticket.
        The caller's tenant is verified against the ticket's TenantID;
        a cross-tenant resolve attempt is indistinguishable from a
        ticket-not-found result — both return 404 with the same
        response body so the endpoint cannot be used to fingerprint
        which ticket IDs belong to another tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalationResolveRequest"
      responses:
        "200":
          description: |
            Resolved. Returns the full updated EscalationTicket so
            clients can render the post-resolve state (outcome,
            resolution notes, resolved_at timestamp, complete
            timeline) without a follow-up GET. The legacy
            EscalationResolveResponse wrapper ({status, ticket_id})
            is retained in this spec for documentation only; the
            handler has always returned a full ticket, so
            generated clients should consume EscalationTicket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationTicket"
        "400":
          description: |
            Client-side input fault. Returned for malformed JSON,
            an unknown outcome value, an oversized request body, or
            a missing ticket_id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: |
            Missing or invalid authentication. Returned before any
            body parsing so an unauthenticated caller cannot probe
            the request schema by observing 400 vs 401 differentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |
            Ticket not found OR the ticket belongs to a different
            tenant. The two cases return the same status code and
            response body so cross-tenant callers cannot fingerprint
            which ticket IDs exist in another tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |
            The ticket has already been resolved. Re-resolving a
            non-pending ticket is a business-rule conflict, not an
            input fault — the response is 409 (canonical "request
            conflicts with current resource state") so clients can
            distinguish a malformed-payload retry from a state-
            conflict requiring a fresh fetch + re-decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: |
            Server-side fault. Returned when a downstream dependency
            fails (Postgres ticket store unreachable, NATS publish
            failure, JSON marshal error). The error body is
            intentionally generic — the diagnostic detail lives in
            the server logs, not on the wire.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/escalation/{ticket_id}:
    get:
      tags: [escalation]
      summary: Fetch an escalation ticket by ID
      description: |
        Returns the structured ticket payload (incident, timeline,
        outcome) for the supplied ticket_id. The caller's tenant is
        verified against the ticket's TenantID — cross-tenant
        accesses return 404, not the ticket, to avoid leaking the
        existence of a ticket from a different tenant.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: string
          description: Ticket identifier returned by the escalation publisher.
      responses:
        "200":
          description: Ticket found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationTicket"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/investigation/message/{pseudo_id}:
    get:
      tags: [investigation]
      summary: Per-message investigation trail
      description: |
        WS-3b operator pivot from an incident / dashboard into the
        message-level verdict context. Returns the persisted
        EvaluationResult for (tenant, pseudo_message_id) plus, when
        the result carries pseudonymised participant hashes, the
        CommunicationHistory row for the (sender, recipient) pair.

        Tenant isolation: cross-tenant accesses return 404 with an
        identical response body to the genuine "not found" case so
        the response surface cannot be used to fingerprint which
        message IDs exist in another tenant. The caller's tenant
        comes from the verified JWT claim — the pseudo_message_id
        path parameter is the lookup key, never the tenant ID.
      parameters:
        - in: path
          name: pseudo_id
          required: true
          schema:
            type: string
          description: |
            Pseudonymised message identifier the evaluator stamped on
            the verdict (BYTEA message_id_hash on
            evaluation_results, rendered as the original wire string
            by the writer). Operators copy this from the incident
            detail view; it is opaque to the SOC.
      responses:
        "200":
          description: Message trail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTrailResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/investigation/sender/{sender_hash}:
    get:
      tags: [investigation]
      summary: Per-sender investigation trail
      description: |
        WS-3b operator pivot from a verdict / dashboard into the
        sender-level activity pattern. Returns the recent
        evaluation verdicts for (tenant, sender_hash), the recipient
        fan-out across the tenant, and a small aggregate envelope
        for the dashboard headline metrics.

        sender_hash is base64url-encoded (no padding) so it fits
        cleanly into a URL path. base64-standard with padding is
        also accepted as a defense-in-depth fallback.

        The look-back is bounded server-side: the optional
        `since_hours` parameter caps at 720 (30 days), matching the
        rolling Count30d window on communication_histories.
      parameters:
        - in: path
          name: sender_hash
          required: true
          schema:
            type: string
          description: |
            base64url-encoded BYTEA pseudonym derived by the bridge /
            enricher (HashPII(tenantID, lower(trim(address)))). The
            dashboard typically copies this from a verdict detail.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
          description: |
            Cap on the number of verdicts returned. Default 100; the
            repository hard-caps at 500 regardless of the requested
            value, but values above 500 are rejected at the handler
            with 400 so the caller can correct the typo.
        - in: query
          name: since_hours
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 720
          description: |
            Aggregation look-back in hours. Default 720 (30 days).
            Values outside [1, 720] return 400.
      responses:
        "200":
          description: Sender trail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SenderTrailResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/dashboard/summary:
    get:
      tags: [dashboard]
      summary: Aggregate dashboard for a tenant
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
        - in: query
          name: range
          schema:
            type: string
            description: "Time range (`7d`, `24h`, etc.)"
            default: "7d"
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "400":
          $ref: "#/components/responses/Error"
  /v1/predict/recipient:
    post:
      tags: [predict]
      summary: Pre-send recipient risk score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictRecipientRequest"
      responses:
        "200":
          description: Prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictRecipientResponse"
  /v1/predict/open:
    post:
      tags: [predict]
      summary: Pre-open URL risk score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictOpenRequest"
      responses:
        "200":
          description: Prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictOpenResponse"
  /v1/onboarding/start:
    get:
      tags: [onboarding]
      summary: Start OAuth consent flow
      description: |
        Generates a redirect URL for the OAuth consent screen. The caller
        should redirect the tenant admin's browser to the returned URL.
      parameters:
        - in: query
          name: provider
          required: true
          schema:
            type: string
            enum: [google_workspace, microsoft_365]
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Redirect URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect_url:
                    type: string
                    format: uri
        "400":
          $ref: "#/components/responses/Error"
...
  /v1/onboarding/status:
    get:
      tags: [onboarding]
      summary: Onboarding status
      description: |
        Returns the current onboarding status for a tenant, including
        user/group discovery counts and the number of users requiring
        admin review.
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Onboarding status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatus"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    BannerActionRequest:
      type: object
      required: [token, action]
      properties:
        token:
          type: string
          description: Signed JWT minted when the banner was rendered.
        action:
          type: string
          enum: [report_phishing, mark_safe, trust_sender, ignore]
    BannerActionResponse:
      type: object
      required: [status, action, pseudonymized_message_id]
      properties:
        status:
          type: string
          example: ok
        action:
          type: string
        pseudonymized_message_id:
          type: string
    QuarantineReleaseRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
        approver_role:
          type: string
          enum: [admin, security_lead, end_user]
        justification:
          type: string
...